admin/requirements.txt
alembic==1.7.5
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
bcrypt==3.2.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
datastore-entity==0.2.0
distlib==0.3.3
dnspython==2.2.0
email-validator==1.1.3
env==0.1.0
filelock==3.4.0
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-CKEditor==0.4.6
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
Flask-WTF==1.0.0
google-api-core==1.31.5
google-auth==1.35.0
google-cloud==0.34.0
google-cloud-core==1.7.2
google-cloud-datastore==2.4.0
google-cloud-storage==1.16.1
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
libcst==0.4.0
Mako==1.1.6
MarkupSafe==2.0.1
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.3
platformdirs==2.4.0
proto-plus==1.19.8
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
rsa==4.8
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
typing-inspect==0.7.1
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.10.0
watchdog==2.1.6
Werkzeug==2.0.2
wtf==0.1
WTForms==3.0.1
